refactor: 💡 change directory of card component
