+++ src/pages/question/card/styled.ts
